Added devcontainer & fixed docker-compose.  Changed app named to be idiomatic.
