--- Capstone-4901---SeaTeam/config/static.yaml
@@ -1,60 +1,0 @@
-admin:
-  address:
-    socket_address: { address: 127.0.0.1, port_value: 9901 }
-
-static_resources:
-  listeners:
-  - name: listener_0
-    address:
-      socket_address: { address: 127.0.0.1, port_value: 10000 }
-    filter_chains:
-    - filters:
-      - name: envoy.filters.network.http_connection_manager
-        typed_config:
-          "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
-          stat_prefix: ingress_http
-          codec_type: AUTO
-          route_config:
-            name: local_route
-            virtual_hosts:
-            - name: local_service
-              domains: ["*"]
-              routes:
-              - match: { prefix: "/" }
-                route: { cluster: some_service }
-          http_filters:
-          - name: envoy.filters.http.router
-            typed_config:
-              "@type": type.googleapis.com/envoy.extensions.filters.http.router.v3.Router
-
-  clusters:
-    - name: some_service
-      connect_timeout: 0.25s
-      type: STATIC
-      lb_policy: LEAST_CONNECTIONS
-      load_assignment:
-        cluster_name: some_service
-        endpoints:
-        - lb_endpoints:
-          - endpoint:
-              address:
-                socket_address:
-                  address: 127.0.0.1
-                  port_value: 1234
-          - endpoint:
-              address:
-                socket_address:
-                  address: 127.0.0.2
-                  port_value: 5678
-          - endpoint:
-              address:
-                socket_address:
-                  address: 127.0.0.3
-                  port_value: 9876
-          - endpoint:
-              address:
-                socket_address:
-                  address: 127.0.0.4
-                  port_value: 5544
-
-
