--- Capstone-4901---SeaTeam/dockerfile
@@ -1,18 +1,0 @@
-FROM golang:1.19.3-buster AS build
-
-WORKDIR /app
-
-COPY go.mod ./
-COPY go.sum ./
-
-RUN go mod download && go mod verify
-
-COPY . .
-
-RUN go build -o /envoyrouter
-
-FROM gcr.io/distroless/base-debian11
-
-COPY --from=build /envoyrouter /envoyrouter
-
-ENTRYPOINT ["/envoyrouter"]
