--- Capstone-4901---SeaTeam/frontend/index.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Envoy Router</title>
-    <link rel="stylesheet" type="text/css" href="styles.css">
-</head>
-<body>
-    <header>
-        <h1>Envoy Router</h1>
-        <p>This is a simple web interface.</p>
-    </header>
-
-    <nav>
-        <ul>
-            <!-- A query variable is passed to each endpoint that is used by determineBackendURL. The actual url that is used is everything before the '?', in this case just '/'-->
-            <li><a href="/?endpoint=0">Go to Endpoint 1</a></li>
-            <li><a href="/?endpoint=1">Go to Endpoint 2</a></li>
-            <li><a href="/?endpoint=2">Go to Endpoint 3</a></li>
-            <li><a href="/health">Health Check</a></li>
-            <li><a href="/endpoint3">Endpoint 3</a></li>
-            <li><a href="/new-feature">New Feature</a></li>
-            <li><a href="/?endpoint=lb">Loadbalancer-decided Endpoint</a></li>
-            <!-- You can add more buttons here -->
-        </ul>
-    </nav>
-</body>
-</html>
