--- Capstone-4901---SeaTeam/frontend/styles.css
@@ -1,52 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #f4f4f4;
-}
-
-header {
-    background-color: #333;
-    color: #fff;
-    padding: 20px 0;
-    text-align: center;
-}
-
-h1 {
-    margin: 0;
-    font-size: 2em;
-}
-
-p {
-    color: #ccc;
-    font-size: 1.2em;
-}
-
-nav {
-    text-align: left;
-    margin: 20px;
-}
-
-ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-li {
-    margin: 10px 0;
-}
-
-a {
-    display: block;
-    padding: 15px 40px;
-    color: #fff;
-    text-decoration: none;
-    background-color: #3498db;
-    border-radius: 8px;
-    transition: background-color 0.3s, color 0.3s;
-}
-
-a:hover {
-    background-color: #2980b9;
-}
