--- Capstone-4901---SeaTeam/go.mod
@@ -1,21 +1,0 @@
-module Capstone-4901---SeaTeam
-
-go 1.21
-
-require (
-	github.com/fsnotify/fsnotify v1.7.0
-	github.com/prometheus/client_golang v1.18.0
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
-	github.com/kr/text v0.2.0 // indirect
-	github.com/matttproud/golang_protobuf_extensions/v2 v2.0.0 // indirect
-	github.com/prometheus/client_model v0.5.0 // indirect
-	github.com/prometheus/common v0.45.0 // indirect
-	github.com/prometheus/procfs v0.12.0 // indirect
-	golang.org/x/sys v0.15.0 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
