+++ docker-compose.yml
@@ -1,31 +1,34 @@
-version: '3.9'
-
 services:
-  envoyrouter:
-    build: ./Capstone-4901---SeaTeam
+  app:
+    build: ./app
     ports:
       - 8000:8000
       - 8081:8081
-
+    volumes:
+      - ./app/config:/config
+  
   prometheus:
-    image: prom/prometheus:v2.40.4
+    image: prom/prometheus:latest
+    volumes:
+      - ./prometheus/prometheus.yml:/etc/prometheus/prometheus.yml:cached
     ports:
       - 9090:9090
-    volumes:
-      - ./prometheus/prometheus.yml:/etc/prometheus/prometheus.yml
-    command:
-      - '--config.file=/etc/prometheus/prometheus.yml'
 
   grafana:
-    image: grafana/grafana:9.3.0
-    ports:
-      - 3000:3000
+    image: grafana/grafana:latest
+    depends_on:
+      - prometheus
     environment:
       - GF_SECURITY_ADMIN_USER=admin
       - GF_SECURITY_ADMIN_PASSWORD=devops123
     volumes:
-      - ./grafana/datasources.yaml:/etc/grafana/provisioning/datasources/datasources.yaml
-      - grafana:/var/lib/grafana
-
-volumes:
-  grafana:
+      - ./grafana/datasources.yaml:/etc/grafana/provisioning/datasources/datasources.yaml:cached
+      - ./grafana/grafana.ini:/etc/grafana/grafana.ini:cached
+    ports:
+      - 3000:3000
+      
+  node-exporter:
+    image: prom/node-exporter:latest
+    restart: unless-stopped
+    expose:
+      - 9100
