+++ grafana/datasources.yaml
@@ -1,7 +1,8 @@
----
 apiVersion: 1
 datasources:
-- name: Main
+- name: Prometheus
   type: prometheus
-  url: http://prometheus:9090
+  url: http://localhost:9090 
   isDefault: true
+  access: proxy
+  editable: true
